requirements.txt
alembic==1.0.8
bcrypt==3.1.6
blinker==1.4
cffi==1.12.2
Click==7.0
coverage==4.5.2
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-Cors==3.0.6
Flask-JWT-Extended==3.13.0
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SocketIO==3.0.2
Flask-SQLAlchemy==2.3.2
gevent==1.3.1
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
monotonic==1.5
nose2==0.7.4
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.9.1
python-editor==1.0.4
python-engineio==3.5.0
python-socketio==4.0.0
six==1.12.0
SQLAlchemy==1.2.18
Werkzeug==0.14.1
